+++ reliably_cli/__main__.py
@@ -3,7 +3,6 @@ from pathlib import Path
 import typer
 
 from .__version__ import __version__
-from .agent.cli import cli as agent_cli
 from .config.cli import cli as config_cli
 from .config.types import Settings
 from .log import console
@@ -11,7 +10,7 @@ from .services.cli import cli as services_cli
 
 cli = typer.Typer()
 cli.add_typer(config_cli, name="config")
-cli.add_typer(agent_cli, name="agent")
+# cli.add_typer(agent_cli, name="agent")
 cli.add_typer(services_cli, name="service")