+++ reliably_cli/config/cli.py
@@ -7,7 +7,7 @@ from rich import print as print_
 
 from . import get_settings
 
-cli = typer.Typer()
+cli = typer.Typer(help="View and read Reliably configuration entries")
 
 
 class EntryType(str, Enum):
@@ -17,7 +17,7 @@ class EntryType(str, Enum):
     host = "host"
 
 
-@cli.command()
+@cli.command(help="Show the current configuration content")
 def view() -> None:
     encoding = "utf-8"
     toml_file = os.getenv("RELIABLY_CLI_CONFIG")
@@ -33,7 +33,7 @@ def view() -> None:
     print_(data, end="")
 
 
-@cli.command()
+@cli.command(help="Read one entry of the configuration")
 def get(entry: EntryType) -> None:
     settings = get_settings()
 
@@ -47,3 +47,33 @@ def get(entry: EntryType) -> None:
             value = settings.service.host
 
     print_(value)
+
+
+@cli.command(help="Create an empty configuration file")
+def create() -> None:
+    toml_file = os.getenv("RELIABLY_CLI_CONFIG")
+    if not toml_file:
+        settings = get_settings()
+        toml_file = settings.__config__.toml_file
+
+    Path(toml_file).touch(exist_ok=True)
+
+
+@cli.command(help="Checks the configuration's presence")
+def exists() -> None:
+    toml_file = os.getenv("RELIABLY_CLI_CONFIG")
+    if not toml_file:
+        settings = get_settings()
+        toml_file = settings.__config__.toml_file
+
+    print("yes" if Path(toml_file).exists() else "no")
+
+
+@cli.command(help="Get the path of the configuration")
+def path() -> str:
+    toml_file = os.getenv("RELIABLY_CLI_CONFIG")
+    if not toml_file:
+        settings = get_settings()
+        toml_file = settings.__config__.toml_file
+
+    print(Path(toml_file).absolute())