+++ reliably_cli/log.py
@@ -1,4 +1,5 @@
 from rich.console import Console
 
-__all__ = ["console"]
+__all__ = ["console", "err_console"]
 console = Console()
+err_console = Console(stderr=True)