+++ reliably_cli/services/cli.py
@@ -2,5 +2,5 @@ import typer
 
 from .plan import cli as plan_cli
 
-cli = typer.Typer()
+cli = typer.Typer(help="Interact with Reliably services")
 cli.add_typer(plan_cli, name="plan")